Protecting API routes

+++ src/app/api/courses/[courseId]/modules/[moduleId]/lessons/[lessonId]/route.ts
@@ -1,4 +1,9 @@
-import { deleteLesson, updateLesson } from "@/backend/controllers/lessonController";
+import {
+  deleteLesson,
+  updateLesson,
+} from "@/backend/controllers/lessonController";
+import { isAdmin } from "@/backend/utils/isAdmin";
+import { NextResponse } from "next/server";
 
 export async function PUT(
   req: Request,
@@ -6,10 +11,12 @@ export async function PUT(
     params,
   }: { params: { courseId: string; moduleId: string; lessonId: string } }
 ) {
-  const { courseId, moduleId, lessonId } = params;
-  const { lesson } = await req.json();
-  updateLesson(courseId, moduleId, lessonId, lesson);
-  return Response.json("");
+  if (await isAdmin()) {
+    const { courseId, moduleId, lessonId } = params;
+    const { lesson } = await req.json();
+    await updateLesson(courseId, moduleId, lessonId, lesson);
+    return NextResponse.json(null);
+  } else return NextResponse.json(null, { status: 401 });
 }
 
 export async function DELETE(
@@ -18,12 +25,13 @@ export async function DELETE(
     params,
   }: { params: { courseId: string; moduleId: string; lessonId: string } }
 ) {
-  const { courseId, moduleId, lessonId } = params;
-  try {
-    await deleteLesson(courseId, moduleId, lessonId);
-    return new Response("", {status: 200})
-  }
-  catch (err) {
-    return new Response(err as string, { status: 400 });
-  }
+  if (await isAdmin()) {
+    const { courseId, moduleId, lessonId } = params;
+    try {
+      await deleteLesson(courseId, moduleId, lessonId);
+      return new NextResponse("", { status: 200 });
+    } catch (err) {
+      return new NextResponse(err as string, { status: 500 });
+    }
+  } else return NextResponse.json(null, { status: 401 });
 }
