+++ src/app/api/courses/[courseId]/modules/[moduleId]/route.ts
@@ -1,31 +1,38 @@
-import { deleteModule, updateModulePatch } from "@/backend/controllers/moduleController";
-import { NextResponse } from "next/server";
+import {
+  deleteModule,
+  updateModulePatch,
+} from "@/backend/controllers/moduleController";
+import { isAdmin } from "@/backend/utils/isAdmin";
+import { NextRequest, NextResponse } from "next/server";
 
 export async function PATCH(
-  req: Request,
+  req: NextRequest,
   { params }: { params: { courseId: string; moduleId: string } }
 ) {
-  const { courseId, moduleId } = params;
-  let { module } = await req.json();
-  try {
-    await updateModulePatch(courseId, moduleId, module)
-    return NextResponse.json("", { status: 200 });
-  }
-  catch (err) {
-    return NextResponse.json("", { status: 400 });
-  }
+  if (await isAdmin()) {
+    const { courseId, moduleId } = params;
+    let { module } = await req.json();
+    try {
+      await updateModulePatch(courseId, moduleId, module);
+      return NextResponse.json(null, { status: 200 });
+    } catch (err) {
+      return NextResponse.json(null, { status: 500 });
+    }
+  } else return NextResponse.json(null, { status: 401 });
 }
 
 export async function DELETE(
-  req: Request,
+  req: NextRequest,
   { params }: { params: { courseId: string; moduleId: string } }
 ) {
-  const { courseId, moduleId } = params;
-  try {
-    await deleteModule(courseId, moduleId);
-    return NextResponse.json("", { status: 200 });
-  } catch (err) {
-    console.log(err);
-    return NextResponse.json(err as string, { status: 400 });
-  }
+  if (await isAdmin()) {
+    const { courseId, moduleId } = params;
+    try {
+      await deleteModule(courseId, moduleId);
+      return NextResponse.json("", { status: 200 });
+    } catch (err) {
+      console.error(err);
+      return NextResponse.json(null, { status: 500 });
+    }
+  } else return NextResponse.json(null, { status: 401 });
 }
