+++ src/app/api/courses/[courseId]/route.ts
@@ -6,6 +6,7 @@ import {
   updateCourse,
 } from "@/backend/controllers/courseController";
 import { APICourseType } from "@/utils/apiTypes";
+import { isAdmin } from "@/backend/utils/isAdmin";
 
 // Load course
 export async function GET(
@@ -28,11 +29,13 @@ export async function POST(
   req: Request,
   { params }: { params: { courseId: string } }
 ) {
-  const courseId = params.courseId;
-  const { title } = await req.json();
-  console.log(`id: ${courseId}`);
-  await createCourse(title, courseId);
-  return NextResponse.json("Suc");
+  if (await isAdmin()) {
+    const courseId = params.courseId;
+    const { title, category } = await req.json();
+    console.log(`id: ${courseId}`);
+    await createCourse(title, courseId, category);
+    return NextResponse.json(null);
+  }
 }
 
 // Update course
@@ -40,27 +43,30 @@ export async function PATCH(
   req: NextRequest,
   { params }: { params: { courseId: string } }
 ) {
-  const courseId = params.courseId;
-  const { course }: { course: APICourseType } = await req.json();
-  try {
-    await updateCourse(courseId, course);
-  } catch (err) {
-    return NextResponse.json({}, { status: 401 });
-  }
+  if (await isAdmin()) {
+    const courseId = params.courseId;
+    const { course }: { course: APICourseType } = await req.json();
+    try {
+      await updateCourse(courseId, course);
+    } catch (err) {
+      return NextResponse.json({}, { status: 500 });
+    }
 
-  return NextResponse.json({}, { status: 200 });
+    return NextResponse.json(null);
+  } else return NextResponse.json(null, { status: 401 });
 }
 
-export async function DELETE(req: NextRequest, { params }: { params: { courseId: string }  }) {
-  const courseId = params.courseId;
-  try {
-    await deleteCourse(courseId);
-  }
-
-  catch (err) {
-    return NextResponse.json({}, { status: 401 });
-  }
-
-  return NextResponse.json({}, { status: 200 });
+export async function DELETE(
+  req: NextRequest,
+  { params }: { params: { courseId: string } }
+) {
+  if (await isAdmin()) {
+    const courseId = params.courseId;
+    try {
+      await deleteCourse(courseId);
+    } catch (err) {
+      return NextResponse.json(null, { status: 500 });
+    }
+    return NextResponse.json(null);
+  } else return NextResponse.json(null, { status: 401 });
 }
-
